refactor(runner): move executor files to dedicated folder
Move executor-related files into lib/executor/:
- executor.ts -> executor/index.ts
- executor-env.ts -> executor/env.ts
- executor-types.ts -> executor/types.ts

Re-export types from index.ts for external consumers.
